tetragon: Move TestKprobeTraceCapabilityChecks to arch specific file
Moving TestKprobeTraceCapabilityChecks to arch specific file,
because it's using amd64 specific syscall.Ioperm syscall.
